requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-jazzmin
psycopg2-binary==2.9.10
redis==5.0.3
django-redis==5.4.0
python-dotenv==1.0.1
Pillow==11.2.1
gunicorn==23.0.0
whitenoise==6.9.0
requests==2.32.3
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
channels[daphne]==4.2.0
channels-redis==4.2.1
# Celery + beat
celery==5.4.0
django-celery-beat==2.7.0
# Cloudinary media storage
cloudinary
django-cloudinary-storage
# PDF generation
weasyprint==62.3
# Optional: SMS via Semaphore (uncomment when needed)
# semaphore-sms==1.0.0
